# CampusOverflow-Front/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - CampusOverflow project","description":"This is a sample OpenApi specification for CampusOverflow project","version":"1.0"},"servers":[{"url":"http://localhost:8000/api/v1","description":"Local ENV"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Report","description":"Report API"},{"name":"Authentication","description":"Authentication API"},{"name":"Question","description":"Question API"},{"name":"Tag","description":"Tag API"}],"paths":{"/tag":{"get":{"tags":["Tag"],"operationId":"findAllTags","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}}}}}}},"put":{"tags":["Tag"],"operationId":"updateTag","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagRequestUpdate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}}},"post":{"tags":["Tag"],"operationId":"saveTag","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagRequestNew"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}}}},"/report":{"put":{"tags":["Report"],"operationId":"updateReport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRequestUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}}}},"post":{"tags":["Report"],"operationId":"saveReport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportRequestNewDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}}}}},"/question":{"get":{"tags":["Question"],"operationId":"findAllQuestions","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseQuestion"}}}}}},"put":{"tags":["Question"],"operationId":"updateQuestion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}}}},"post":{"tags":["Question"],"operationId":"saveQuestion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}}}}},"/vote":{"post":{"tags":["vote-controller"],"operationId":"saveVote","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoteRequestNewDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vote"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/auth/activate-account":{"post":{"tags":["Authentication"],"operationId":"activateAccount","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/answer":{"post":{"tags":["answer-controller"],"operationId":"saveAnswer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerRequestNewDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Answer"}}}}}},"patch":{"tags":["answer-controller"],"operationId":"updateAnswer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerRequestUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Answer"}}}}}}},"/vote/{id}":{"get":{"tags":["vote-controller"],"operationId":"findVoteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vote"}}}}}},"delete":{"tags":["vote-controller"],"operationId":"deleteVoteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/vote/user/{userId}":{"get":{"tags":["vote-controller"],"operationId":"findAllVotesByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/vote/answer/{answerId}":{"get":{"tags":["vote-controller"],"operationId":"findAllVotesByAnswerId","parameters":[{"name":"answerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/tag/{id}":{"get":{"tags":["Tag"],"operationId":"findTagById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}}},"delete":{"tags":["Tag"],"operationId":"deleteTagById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/tag/search":{"get":{"tags":["Tag"],"operationId":"searchTags","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}}}}}}}},"/tag/name/{name}":{"get":{"tags":["Tag"],"operationId":"findTagByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}}}},"/report/{id}":{"get":{"tags":["Report"],"operationId":"findReportById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}}}},"delete":{"tags":["Report"],"operationId":"deleteReportById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/report/user/{userId}":{"get":{"tags":["Report"],"operationId":"findAllReportsByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/report/search":{"get":{"tags":["Report"],"operationId":"searchReports","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/question/{id}":{"get":{"tags":["Question"],"operationId":"findQuestionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}}}},"delete":{"tags":["Question"],"operationId":"deleteQuestionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/question/user":{"get":{"tags":["Question"],"operationId":"findAllQuestionsByUser","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseQuestion"}}}}}}},"/question/search":{"get":{"tags":["Question"],"operationId":"searchQuestions","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"query","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseQuestion"}}}}}}},"/answer/{id}":{"get":{"tags":["answer-controller"],"operationId":"findAnswerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Answer"}}}}}},"delete":{"tags":["answer-controller"],"operationId":"deleteAnswerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/answer/user/{userId}":{"get":{"tags":["answer-controller"],"operationId":"findAllAnswersByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/answer/question/{questionId}":{"get":{"tags":["answer-controller"],"operationId":"findAllAnswersByQuestionId","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"TagRequestUpdate":{"required":["id","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"Tag":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"integer","format":"int64"},"updatedBy":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ReportRequestUpdateDto":{"required":["description","id","title","type","userId"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["QUESTION_REPORT","ANSWER_REPORT","USER_REPORT"]},"reportedUserId":{"type":"integer","format":"int64"},"reportedAnswerId":{"type":"integer","format":"int64"},"reportedQuestionId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"resolved":{"type":"boolean"}}},"Answer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"integer","format":"int64"},"updatedBy":{"type":"integer","format":"int64"},"content":{"type":"string"},"user":{"$ref":"#/components/schemas/User"},"question":{"$ref":"#/components/schemas/Question"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Question":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"integer","format":"int64"},"updatedBy":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"user":{"$ref":"#/components/schemas/User"}}},"Report":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"integer","format":"int64"},"updatedBy":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["QUESTION_REPORT","ANSWER_REPORT","USER_REPORT"]},"reportedUser":{"$ref":"#/components/schemas/User"},"reportedAnswer":{"$ref":"#/components/schemas/Answer"},"reportedQuestion":{"$ref":"#/components/schemas/Question"},"user":{"$ref":"#/components/schemas/User"},"resolved":{"type":"boolean"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"birthdate":{"type":"string","format":"date"},"role":{"type":"string","enum":["USER","ADMIN"]},"enabled":{"type":"boolean"},"accountLocked":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonLocked":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"}}},"VoteRequestNewDto":{"required":["answerId","type","userId"],"type":"object","properties":{"answerId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["UPVOTE","DOWNVOTE"]}}},"Vote":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"integer","format":"int64"},"updatedBy":{"type":"integer","format":"int64"},"answer":{"$ref":"#/components/schemas/Answer"},"user":{"$ref":"#/components/schemas/User"},"type":{"type":"string","enum":["UPVOTE","DOWNVOTE"]}}},"TagRequestNew":{"required":["name"],"type":"object","properties":{"name":{"type":"string"}}},"ReportRequestNewDto":{"required":["description","title","type","userId"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["QUESTION_REPORT","ANSWER_REPORT","USER_REPORT"]},"reportedUserId":{"type":"integer","format":"int64"},"reportedAnswerId":{"type":"integer","format":"int64"},"reportedQuestionId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"resolved":{"type":"boolean"}}},"QuestionRequest":{"required":["content","tags","title"],"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"tags":{"type":"array","items":{"type":"integer","format":"int64"}}}},"RegistrationRequest":{"required":["email","name","password","username"],"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"birthdate":{"type":"string","format":"date"}}},"LoginRequest":{"required":["password","usernameOrEmail"],"type":"object","properties":{"usernameOrEmail":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"}}},"AnswerRequestNewDto":{"required":["content","questionId","userId"],"type":"object","properties":{"content":{"type":"string"},"userId":{"type":"integer","format":"int64"},"questionId":{"type":"integer","format":"int64"}}},"AnswerRequestUpdateDto":{"required":["content","id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"}}},"PageResponseQuestion":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/Question"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}